+++ Changelog.md
@@ -41,6 +41,7 @@ This is an attempt to document all the changes that are part of this.
 - moved `utils.simple_player_name` to `utils.sanitise.simple_player_name`.
 - moved functionality of `get_player_by_role` into the Manager.
 - removed unused `discord_id` property from `Player`.
+- moved `get_keywords`, `_manage_coast_signature`, `get_unit_type`, `parse_season` & `get_value_from_timestamp` to `sanitise.py`
 
 ## Other Changes
 - renamed ordered scoreboard functions.